feat: Profiler in portable executor (#1004)
We used to have profiler in SP1 v5, but MinimalExecutor ships without profiler.
This changes adds profiler to the portable executor back. It also does a small
refactor: previously, an ELF is parsed by 2 crates twice, once for the main loader,
once for profiler. This change fixes the code so an ELF is only parsed once.