feat: New way to make max memory / input region size tweakable (#1002)
* feat: New way to make max memory / input region size tweakable

The same thing was implemented in #962 with a different design. While #962
relies on Rust macro parameters, this change uses environment
variables at build time. The problem with #962, is that we cannot save
max memory in a static constant from ELF note section. This newer design
works around the problem.

In addition, the newer design is also simpler than previous one.

* fix: cargo fmt